tools: derive known arguments from the generated schema
Review feedback: validate request arguments against the schema built by
createJSONSchemaFromHandler — the same one advertised to clients —
instead of a parallel reflection helper. Validation helpers move to
tools_param_validation.go; the print-friendly argument list is computed
only in the error path.
